Add support for multiple module decorators and providers
Support multiple module decorators and add "module private providers"
to provide objects only within module's scope.

This is primarily needed for providing a "job.Group" for each module without
creating the group and adding it to the lifecycle if it is not used.

Add a test case to job package to show how this is used.

Signed-off-by: Jussi Maki <jussi@isovalent.com>