hive: take Logger as part of Start/Stop or Run
Instead of having to provide a logger already at hive creation time,
pass the logger to Start/Stop/Run, which is when the underlying cells
actually need it.

Also revert the changes to the Apply interface by passing the logger
when the invokes are actually run instead of during apply.

Finally, add hivetest.Logger to create a logger from a testing.TB, so
that the logs can be attributed to the correct test when running tests
in parallel. Unfortunately, t.Helper() doesn't correctly work in
combination with slog - we log the real source location as an attribute
as a workaround.

Signed-off-by: David Bimmler <david.bimmler@isovalent.com>