hivetest/slog: don't log after the test is done
Since the logger reference may be passed to some background thread which
runs concurrently to the main test, it's possible that attempts to log
occur after the test has finished. Go's testing package chooses to panic
in that circumstance, hence we should avoid it.

Notably, it's probably a context bug of calling code to log after the
test completes, but if we don't try to work around this at this level,
the transition from logrus to slog will be painful, since some tests
would just non-deterministically panic.

Signed-off-by: David Bimmler <david.bimmler@isovalent.com>