job: Add WithDebounce option to trigger
Add WithDebounce functional option to the timer job trigger.  This
allows to fold multiple triggers over the specified interval.  Also, it
is now possible to have a timer job with a zero interval, that runs its
function only when explicitly triggered.

This replicates the behavior of the github.com/cilium/cilium/pkg/trigger
package in the Cilium codebase, making the same semantic available
within a job from the hive framework.

Note that the debounce is striclty tied to the trigger and does not
affect the behavior of the timer job when the internal timer expires. In
that case, the job is run without taking into account the debounce
interval.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>