job: Sanitize instead of validate the job names
The job names may come from many different sources and
it is too much of a foot-gun to reject the names by panicing.

Instead to keep the names within a nice constraint set of
characters and length, sanitize the names silently.

Related PR that had to work around the panics:
https://github.com/cilium/cilium/pull/35031

Signed-off-by: Jussi Maki <jussi.maki@isovalent.com>