script: Vendor in rsc.io/script
Vendor in the rsc.io/script library. As it was essentially
a code dump out from go's internals and is unmaintained by
rsc, it's easiest to vendor it here.

It is vendored into cilium/hive instead of its own repo as
its use and extensions will be somewhat coupled with hive.

As first extension, add support for interactive use that
doesn't wait for the section end.

Signed-off-by: Jussi Maki <jussi.maki@isovalent.com>