script: Add support for interrupting long-running commands
In order to implement e.g. "watch" command for StateDB we need
to be able to interrupt the command. To do this, we need to be able
to catch SIGINT, which requires the ISIG lflag and for that we'll
need our own "MakeRaw" function.

Additionally to be able to dump the log buffer periodically we
add a FlushLog() method.

Signed-off-by: Jussi Maki <jussi.maki@isovalent.com>