script: Add '*' prefix for retrying commands
This adds a generic facility for retrying failing commands.

A command that is prefixed with '*' and fails will cause the whole
section (delimited by '#' comments) to be retried from the top.

Retrying is repeated until the command succeeds or the context is
cancelled. The retry interval can be set in (*Engine).RetryInterval.

Example usage:

  # Verify table contents
  db show -o=out.table my-table
  * cmp out.table expected.table

Signed-off-by: Jussi Maki <jussi.maki@isovalent.com>