scripttest: Add support for shebang args
Sometimes it's useful to change the test parameters from the script
itself.

To allow for that, add support for parsing the first line of the script
as arguments and pass those to the engine constructor. For example the
test script:

  #! -foo=bar -baz=1
  ...

will cause the engine constructor to be called with
[]string{"-foo=bar", "-baz=1"}. These can then be parsed as one wants,
though with the idea that they could be used to configure the hive:

  func myTestEngine(t testing.TB, args []string) *scripttest.Engine {
    h := hive.New(...)
    flags := pflag.NewFlagSet("", pflag.ContinueOnError)
    h.RegisterFlags(flags)
	flags.Parse(args)
	...
  }

Signed-off-by: Jussi Maki <jussi.maki@isovalent.com>