job: handle ctx cancellation in complete function
Currently, the termination of an observer job due to context
cancellation is logged as error. This might be confusing and lead
to unnecessary noise (e.g. in case of graceful shutdown).

The next function already handles the context cancellation differently
too.

Therefore, this commit only logs the error, if it's not of type
`context.Canceled`.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>