job: Assign scoped health to jobs added to scoped job groups
Jobs can be added to a job group both before and after it has started.
When adding a job group before it has started it is put in a queue and
started on group start.

One of the parameters we pass in when the job starts is a health
reporter. However, there was a bug where when you create a scoped job
group, we would pass the health reporter of its parent to the job and
not the health reporter of the scoped job group. We were doing it
correctly for jobs added to the job group after it had started.

This commit fixes the bug enqueueing the health reporter along with the
job when enqueueing before the job group has started and using the
colocated health reporter to call the job on start.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>