Add AppendWithModuleID to DefaultLifecycle
In cilium/cilium the job groups were provided for each module and its
lifecycle hooks appended when constructed. This is problematic when the
job group is used multiple times as then the dependencies of the later
uses won't influence the order in which start hooks run, e.g. if a later
job depends on X it might happen that the job is started before X is started.

To fix this, we'll make the job groups start as the very last thing in the
lifecycle which ensures all the start hooks of provided objects are run.

For that we'll need to use a separate lifecycle for the job groups and to
keep the output nice we'll need to be able to provide the module ID when
appending the job group hooks. This commit exposes 'AppendWithModuleID'
which we can use to do that.

Signed-off-by: Jussi Maki <jussi.maki@isovalent.com>