script: Refuse empty files in cmp when updating
Add an additional safeguard to -scripttest.update when used with 'cmp'
to ignore empty files and increase the retries to 2. This significantly
increases the likelyhood that we do not pick a too "early" version of the
actual output.

Removed the update handling from 'empty' command as I don't see a use-case
where we'd use 'empty' on a file within the txtar. You can just empty the
file manually.

Signed-off-by: Jussi Maki <jussi.maki@isovalent.com>