job: Append jobs added before starting to lifecycle
In cilium/cilium the job groups were provided for each module and its
lifecycle hooks appended when constructed. This is problematic when the
job group is used multiple times as then the dependencies of the later
uses won't influence the order in which start hooks run, e.g. if a later
job depends on X it might happen that the job is started before X is started.

For example:

  cell.Module(...,
    cell.Provide(newX),

    // First use of the job.Group will construct it and append hooks to
    // lifecycle.
    cell.Invoke(func(jg job.Group, ...) { jg.Add( ... )}),

    // This second use will construct X and append its hooks, but these
    // are now *after* the job.Group's hooks and hence we might start the
    // job before X has been started.
    cell.Invoke(func(jg job.Group, x *X) { jg.Add( ... )}),
  )

To fix this, let's use the lifecycle to start the jobs that are added
before the job group is started. This way the start order will respect
the order in which [job.Group.Add] is called and thus be arranged according
to the dependency order.

This is a breaking change in the interface, but there's not that many
direct uses of "NewGroup" in cilium/cilium and thus fairly easy to fix.

Signed-off-by: Jussi Maki <jussi.maki@isovalent.com>