scripttest: Add NoParallel option
We still have global variables in cilium/cilium and tests that end up
touching those will need to be run sequentially. To make this easier
to orchestrate without having to e.g. use a mutex inside 'newEngine'
add the 'NoParallel' option.

We are not making parallelism opt-in as we really want this to be the
default way we run tests.

Signed-off-by: Jussi Maki <jussi.maki@isovalent.com>