script: Add tab completion for flags and arguments
This commit adds tab completion for the shell. When in the REPL
the <tab> key can be used to give suggestions. When typing the initial
command it will suggest commands starting with the typed characters.
Repeated presses of the <tab> key will cycle through the suggestions.

When pressing <tab> while in an argument or flag value, a callback
on the command is called if it exists. This callback can return a
slice of values to suggest. For arguments, all arguments up and
including the current argument are passed to the callback. For flags,
the flag name and value typed so far are passed to the callback.

To accommodate the "remote shell" used by Cilium, a hidden flag is now
implicitly added to all commands `--autocomplete`. When a command
is called with this flag the engine will intercept the command
and return a line separated list of suggestions instead of executing
the command. For example: `test --autocomplete='arg1 --flag1=val1 ar'`
will be interpreted the same as `test arg1 --flag1=val1 ar` would on
the REPL and the output will be a list of suggestions:

```
> test --autocomplete='arg1 --flag1=val1 ar'
test arg1 --flag1=val1 argument2
test arg1 --flag1=val1 arg2
test arg1 --flag1=val1 args2
```

The `help` command also has a hidden flag `-a` which if passed will
simply list all available commands without additional help text.

```
> help -a
cat
cd
chmod
```

This should allow the remote shell in Cilium to provide tab completion
in the same way as the REPL.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>