script: Allow setting env var value from stdout
This is useful if there are commands that output values that could be
used as parameters to subsequent commands.

Also modify the scripttest runner when verbose mode is set to execute
the "env" command using ExecuteLine() instead of running Env() directly
because we now need the flags to be initialized.

Co-authored-by: Fabio Falzoi <fabio.falzoi@isovalent.com>
Signed-off-by: Chris Tarazi <chris@isovalent.com>