shell: Propagate failure in ShellExchange
If the script command fails propagate the status with
an error marker that is sent instead of the command end
marker and use this to return an error in ShellExchange.

With this the non-interactive use of the shell will exit
with non-zero exit status on command failure, e.g.
"cilium-agent shell -- unknown command" will exit with status 1.

Signed-off-by: Vipul Singh <singhvipul@microsoft.com>
Signed-off-by: Jussi Maki <jussi@isovalent.com>